+++ CHANGELOG.md
@@ -38,6 +38,7 @@
 * [#270] [I18n] Don't apologize about missing relationship support.
 * [#237] [I18n] Fix broken paths for several I18n files (de, es, fr, pt-BR, vi).
 * [#266] [OPTIM] Save a few database queries by using cached counts
+* [#295] [FEATURE] Add dashboard detection for ActiveRecord::Enum fields.
 
 ### 0.1.1 (November 12, 2015)